ivar2: Use connection.uri
This change makes it possible to specify tcpv6 and tls in connection.
Not that I've had some problems with binding laddr, so I removed the
example for now.

This change also improves the hostmask, instead of relying on user to
specify correct hostmask we parse the incoming hostmask from the server
on JOINs.


Former-commit-id: e2eb5759fa6a91d8a9ae8ca3de5c5e5d7063249e [formerly 49afdf3c7aae6a8208a6117084cc18dbd613a3b4]
Former-commit-id: bbc46433c7ccd87979c7578ff8458ad27bac25d1