ivar2: Use connection.uri
This change makes it possible to specify tcpv6 and tls in connection.
Not that I've had some problems with binding laddr, so I removed the
example for now.

This change also improves the hostmask, instead of relying on user to
specify correct hostmask we parse the incoming hostmask from the server
on JOINs.

Conflicts:
	ivar2.lua

Former-commit-id: 782b4fe22e409a441da137de59b982b7045c08b4 [formerly 485e6754076f771b65bd65d213b7757c120b63d4]
Former-commit-id: 8d4f39925b663cc893634a19d006dbef223c046c